Updated: mudando os tipos

+++ src/main/java/com/example/demo/core/service/Produto/DeletarProdutoService.java
@@ -21,8 +21,14 @@ public class DeletarProdutoService implements DeletarProdutosUseCase {
         Produto produto = produtoRepository.findByCodigoEAN13(codigoEAN_13);
         Estoque estoque = estoqueRepository.findByProduto(produto);
 
-        System.out.println(produto.getId());
-        estoqueRepository.delete(estoque);
-        produtoRepository.delete(produto);
+        if(produto != null && estoque != null){
+            System.out.println(produto.getId());
+            estoqueRepository.delete(estoque);
+            produtoRepository.delete(produto);
+        } else {
+            System.out.println("Esse produto nn existe");
+            throw new Error("Esse produto nn existe");
+        }
+
     }
 }
